requirements.txt
kiteconnect>=4.1.0
pandas>=2.0.0
numpy>=1.24.0
scipy
openpyxl
xgboost>=1.7.0
scikit-learn>=1.3.0

# PostgreSQL
psycopg2-binary>=2.9.0

# Environment config
python-dotenv>=1.0.0

# Automated login
selenium>=4.0.0
webdriver-manager>=4.0.0
pyotp>=2.9.0
